+++ resources/history.txt
@@ -1,0 +1,5 @@
+User: Erik Wallenius.
+Activity: Deposit 1000 kr.
+Time: 14:30:56.
+Date: 2022-12-16
+
