+++ src/Client/Client.java
@@ -2,7 +2,7 @@ package Client;
 
 import Database.Database;
 import Users.Customer;
-import utility.Utility;
+import Utility.Utility;
 
 import java.time.LocalDate;
 import java.util.Scanner;
@@ -70,13 +70,11 @@ public class Client {
         boolean startLoop = true;
         do {
             int answer = utility.inputInt("Välkommen " + customer.getName() +
-                    "\n1. Överföra pengar\n2. Sätta in pengar\n3.\n4. Logga ut");
+                    "\n1. Överföra pengar\n2. Sätta in pengar\n3. Kolla Kontot\n4. Logga ut");
                 switch (answer) {
-                    case (1) -> utility.transfer(customer, database.getCustomers());
-                    case (2) -> {
-                    }
-                    case (3) -> {
-                    }
+                    //case (1) -> utility.transfer(customer, database.getCustomers());
+                    case (2) -> utility.deposit(customer);
+                    case (3) -> utility.checkAccount(customer);
                     case (4) -> startLoop = false;
                     default -> System.out.println("Felaktigt nummer");
                 }
